services/api/pyproject.toml
[project]
name = "tkp-api"
version = "0.1.0"
description = "API service for multi-tenant knowledge platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.11.0",
    "pyjwt[crypto]>=2.10.1",
    "redis>=5.0.8",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tkp_api"]
